- Сделал более читаемый код.
